+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Add option to view the latest changelog.
 - Auto show changelog when a new version is installed.
 - Add `Liked photos` dialog to view all your liked photos.
+- Fix version crash.
+- Improve performance for dropdowns.
 
 # 0.1.1
 
@@ -10,6 +12,6 @@
 - Make weekly refresh rate update background every Monday at midnight.
 - Fix scrollbars.
 
-# 0.1.0 
+# 0.1.0
 
 - Initial release